@@ -6,10 +6,10 @@ export const globalTypes = {
       icon: 'circlehollow',
       title: 'Theme',
       items: [
-        { value: 'light', icon: 'circlehollow', title: 'light' },
-        { value: 'dark', icon: 'circle', title: 'dark' },
-        { value: 'side-by-side', icon: 'sidebar', title: 'side by side' },
-        { value: 'stacked', icon: 'bottombar', title: 'stacked' },
+        { value: 'light', icon: 'sun', title: 'light' },
+        { value: 'dark', icon: 'moon', title: 'dark' },
+        { value: 'side-by-side', icon: 'sidebyside', title: 'side by side' },
+        { value: 'stacked', icon: 'stacked', title: 'stacked' },
       ],
     },
   },
