@@ -12,6 +12,7 @@ export class RendererFactory {
     storyId: string,
     targetDOMNode: HTMLElement
   ): Promise<AbstractRenderer | null> {
+    const targetId = targetDOMNode.id;
     // do nothing if the target node is null
     // fix a problem when the docs asks 2 times the same component at the same time
     // the 1st targetDOMNode of the 1st requested rendering becomes null 🤷‍♂️
@@ -22,17 +23,17 @@ export class RendererFactory {
     const renderType = getRenderType(targetDOMNode);
     // keep only instances of the same type
     if (this.lastRenderType && this.lastRenderType !== renderType) {
-      await AbstractRenderer.resetPlatformBrowserDynamic();
+      await AbstractRenderer.resetApplications();
       clearRootHTMLElement(renderType);
       this.rendererMap.clear();
     }
 
-    if (!this.rendererMap.has(storyId)) {
-      this.rendererMap.set(storyId, this.buildRenderer(storyId, renderType));
+    if (!this.rendererMap.has(targetId)) {
+      this.rendererMap.set(targetId, this.buildRenderer(storyId, renderType));
     }
 
     this.lastRenderType = renderType;
-    return this.rendererMap.get(storyId);
+    return this.rendererMap.get(targetId);
   }
 
   private buildRenderer(storyId: string, renderType: RenderType) {
