@@ -38,7 +38,7 @@ export const mainjsFramework: Fix<MainjsFrameworkRunOptions> = {
     }
 
     const main = await readConfig(mainConfig);
-    const currentFramework = main.getFieldValue(['framework']);
+    const currentFramework = main.getFieldNode(['framework']);
     const features = main.getFieldValue(['features']);
 
     if (currentFramework) return null;
