@@ -5,6 +5,7 @@ import { getStorybookInfo } from '@storybook/core-common';
 import { loadPartialConfigAsync } from '@babel/core';
 import { readConfig } from '@storybook/csf-tools';
 import type { Fix } from '../types';
+import { generateStorybookBabelConfigInCWD } from '../../babel-config';
 
 interface MissingBabelRcOptions {
   needsBabelRc: boolean;
@@ -16,13 +17,11 @@ const frameworksThatNeedBabelConfig = [
   '@storybook/react-webpack5',
   '@storybook/vue-webpack5',
   '@storybook/vue3-webpack5',
-  '@storybook/preact-webpack5',
   '@storybook/html-webpack5',
 ];
 
 export const missingBabelRc: Fix<MissingBabelRcOptions> = {
   id: 'missing-babelrc',
-  promptOnly: true,
 
   async check({ packageManager }) {
     const packageJson = packageManager.retrievePackageJson();
@@ -47,20 +46,21 @@ export const missingBabelRc: Fix<MissingBabelRcOptions> = {
 
     const main = await readConfig(mainConfig);
 
-    const frameworkField = main.getFieldValue(['framework']);
-    const frameworkPackage =
-      typeof frameworkField === 'string' ? frameworkField : frameworkField?.name;
+    const frameworkPackage = main.getNameFromPath(['framework']);
 
-    const addons: any[] = main.getFieldValue(['addons']) || [];
+    const addons = main.getNamesFromPath(['addons']);
 
-    const hasCraPreset = addons.find((addon) => {
-      const name = typeof addon === 'string' ? addon : addon.name;
-      return name === '@storybook/preset-create-react-app';
-    });
+    const hasCraPreset =
+      addons && addons.find((addon) => addon === '@storybook/preset-create-react-app');
 
-    if (frameworksThatNeedBabelConfig.includes(frameworkPackage) && !hasCraPreset) {
+    if (
+      frameworkPackage &&
+      frameworksThatNeedBabelConfig.includes(frameworkPackage) &&
+      !hasCraPreset
+    ) {
       const config = await loadPartialConfigAsync({
         babelrc: true,
+        filename: '__fake__.js', // somehow needed to detect .babelrc.* files
       });
 
       if (!config.config && !config.babelrc && !packageJson.babel) {
@@ -72,19 +72,15 @@ export const missingBabelRc: Fix<MissingBabelRcOptions> = {
   },
   prompt() {
     return dedent`
-      ${chalk.bold(
-        chalk.red('Attention')
-      )}: We could not automatically make this change. You'll need to do it manually.
-
       We detected that your project does not have a babel configuration (.babelrc, babel.config.js, etc.).
 
-      In version 6.x, Storybook provided its own babel settings out of the box. Now, Storybook re-uses your project's babel configuration, with small, incremental updates from Storybook addons.
+      In version 6.x, Storybook provided its own babel settings out of the box. Now, Storybook re-uses ${chalk.bold(
+        "your project's babel configuration"
+      )}, with small, incremental updates from Storybook addons.
 
-      If your project does not have a babel configuration file, you can generate one that's equivalent to the 6.x defaults with the following command in your project directory:
+      If your project does not have a babel configuration file, we can generate one that's equivalent to the 6.x defaults for you. Keep in mind that this can affect your project if it uses babel, and you may need to make additional changes based on your projects needs.
 
-      ${chalk.blue('npx storybook@next babelrc')}
-
-      This will create a ${chalk.blue(
+      We can create a ${chalk.blue(
         '.babelrc.json'
       )} file with some basic configuration and add any necessary package devDependencies.
 
@@ -93,5 +89,9 @@ export const missingBabelRc: Fix<MissingBabelRcOptions> = {
         'https://github.com/storybookjs/storybook/blob/next/MIGRATION.md#babel-mode-v7-exclusively'
       )}
     `;
+  },
+  async run() {
+    logger.info();
+    await generateStorybookBabelConfigInCWD();
   },
 };
