@@ -1,12 +1,17 @@
 import { dirname } from 'path';
 import type { SupportedFrameworks, SupportedRenderers } from './project_types';
+import { externalFrameworks } from './project_types';
 
 export function getCliDir() {
   return dirname(require.resolve('@storybook/cli/package.json'));
 }
 
 export function getRendererDir(renderer: SupportedFrameworks | SupportedRenderers) {
+  const externalFramework = externalFrameworks.find((framework) => framework.name === renderer);
+  const frameworkPackageName = externalFramework?.packageName ?? `@storybook/${renderer}`;
   return dirname(
-    require.resolve(`@storybook/${renderer}/package.json`, { paths: [process.cwd()] })
+    require.resolve(`${frameworkPackageName}/package.json`, {
+      paths: [process.cwd()],
+    })
   );
 }
