@@ -15,8 +15,13 @@ function eqMajor(versionRange: string, major: number) {
   return validRange(versionRange) && minVersion(versionRange).major === major;
 }
 
+/** A list of all frameworks that are supported, but use a package outside the storybook monorepo */
+export const externalFrameworks: { name: SupportedFrameworks; packageName: string }[] = [
+  { name: 'qwik', packageName: 'storybook-framework-qwik' },
+];
+
 // Should match @storybook/<framework>
-export type SupportedFrameworks = 'nextjs' | 'angular' | 'sveltekit';
+export type SupportedFrameworks = 'nextjs' | 'angular' | 'sveltekit' | 'qwik';
 
 // Should match @storybook/<renderer>
 export type SupportedRenderers =
@@ -32,6 +37,7 @@ export type SupportedRenderers =
   | 'marko'
   | 'preact'
   | 'svelte'
+  | 'qwik'
   | 'rax'
   | 'aurelia'
   | 'html'
@@ -51,6 +57,7 @@ export const SUPPORTED_RENDERERS: SupportedRenderers[] = [
   'marko',
   'preact',
   'svelte',
+  'qwik',
   'rax',
   'aurelia',
 ];
@@ -69,12 +76,12 @@ export enum ProjectType {
   SFC_VUE = 'SFC_VUE',
   ANGULAR = 'ANGULAR',
   EMBER = 'EMBER',
-  ALREADY_HAS_STORYBOOK = 'ALREADY_HAS_STORYBOOK',
   WEB_COMPONENTS = 'WEB_COMPONENTS',
   MITHRIL = 'MITHRIL',
   MARIONETTE = 'MARIONETTE',
   MARKO = 'MARKO',
   HTML = 'HTML',
+  QWIK = 'QWIK',
   RIOT = 'RIOT',
   PREACT = 'PREACT',
   SVELTE = 'SVELTE',
@@ -82,6 +89,7 @@ export enum ProjectType {
   RAX = 'RAX',
   AURELIA = 'AURELIA',
   SERVER = 'SERVER',
+  NX = 'NX',
 }
 
 export enum CoreBuilder {
@@ -169,6 +177,13 @@ export const supportedTemplates: TemplateConfiguration[] = [
     },
   },
   {
+    preset: ProjectType.QWIK,
+    dependencies: ['@builder.io/qwik'],
+    matcherFunction: ({ dependencies }) => {
+      return dependencies.every(Boolean);
+    },
+  },
+  {
     preset: ProjectType.REACT_PROJECT,
     peerDependencies: ['react'],
     matcherFunction: ({ peerDependencies }) => {
@@ -190,20 +205,6 @@ export const supportedTemplates: TemplateConfiguration[] = [
     dependencies: ['react-scripts'],
     matcherFunction: ({ dependencies, files }) => {
       return dependencies.every(Boolean) || files.every(Boolean);
-    },
-  },
-  {
-    preset: ProjectType.WEBPACK_REACT,
-    dependencies: ['react', 'webpack'],
-    matcherFunction: ({ dependencies }) => {
-      return dependencies.every(Boolean);
-    },
-  },
-  {
-    preset: ProjectType.REACT,
-    dependencies: ['react'],
-    matcherFunction: ({ dependencies }) => {
-      return dependencies.every(Boolean);
     },
   },
   {
@@ -284,6 +285,22 @@ export const supportedTemplates: TemplateConfiguration[] = [
       return dependencies.every(Boolean);
     },
   },
+  // DO NOT MOVE ANY TEMPLATES BELOW THIS LINE
+  // React is part of every Template, after Storybook is initialized once
+  {
+    preset: ProjectType.WEBPACK_REACT,
+    dependencies: ['react', 'webpack'],
+    matcherFunction: ({ dependencies }) => {
+      return dependencies.every(Boolean);
+    },
+  },
+  {
+    preset: ProjectType.REACT,
+    dependencies: ['react'],
+    matcherFunction: ({ dependencies }) => {
+      return dependencies.every(Boolean);
+    },
+  },
 ];
 
 // A TemplateConfiguration that matches unsupported frameworks
@@ -300,11 +317,7 @@ export const unsupportedTemplate: TemplateConfiguration = {
   },
 };
 
-const notInstallableProjectTypes: ProjectType[] = [
-  ProjectType.UNDETECTED,
-  ProjectType.UNSUPPORTED,
-  ProjectType.ALREADY_HAS_STORYBOOK,
-];
+const notInstallableProjectTypes: ProjectType[] = [ProjectType.UNDETECTED, ProjectType.UNSUPPORTED];
 
 export const installableProjectTypes = Object.values(ProjectType)
   .filter((type) => !notInstallableProjectTypes.includes(type))
