@@ -1,90 +1,90 @@
 // auto generated file, do not edit
 export default {
-  '@storybook/addon-a11y': '7.0.0-beta.29',
-  '@storybook/addon-actions': '7.0.0-beta.29',
-  '@storybook/addon-backgrounds': '7.0.0-beta.29',
-  '@storybook/addon-controls': '7.0.0-beta.29',
-  '@storybook/addon-docs': '7.0.0-beta.29',
-  '@storybook/addon-essentials': '7.0.0-beta.29',
-  '@storybook/addon-highlight': '7.0.0-beta.29',
-  '@storybook/addon-interactions': '7.0.0-beta.29',
-  '@storybook/addon-jest': '7.0.0-beta.29',
-  '@storybook/addon-links': '7.0.0-beta.29',
-  '@storybook/addon-measure': '7.0.0-beta.29',
-  '@storybook/addon-outline': '7.0.0-beta.29',
-  '@storybook/addon-storyshots': '7.0.0-beta.29',
-  '@storybook/addon-storyshots-puppeteer': '7.0.0-beta.29',
-  '@storybook/addon-storysource': '7.0.0-beta.29',
-  '@storybook/addon-toolbars': '7.0.0-beta.29',
-  '@storybook/addon-viewport': '7.0.0-beta.29',
-  '@storybook/addons': '7.0.0-beta.29',
-  '@storybook/angular': '7.0.0-beta.29',
-  '@storybook/manager-api': '7.0.0-beta.29',
-  '@storybook/blocks': '7.0.0-beta.29',
-  '@storybook/builder-manager': '7.0.0-beta.29',
-  '@storybook/builder-vite': '7.0.0-beta.29',
-  '@storybook/builder-webpack5': '7.0.0-beta.29',
-  '@storybook/channel-postmessage': '7.0.0-beta.29',
-  '@storybook/channel-websocket': '7.0.0-beta.29',
-  '@storybook/channels': '7.0.0-beta.29',
-  '@storybook/cli': '7.0.0-beta.29',
-  '@storybook/client-api': '7.0.0-beta.29',
-  '@storybook/client-logger': '7.0.0-beta.29',
-  '@storybook/codemod': '7.0.0-beta.29',
-  '@storybook/components': '7.0.0-beta.29',
-  '@storybook/core-client': '7.0.0-beta.29',
-  '@storybook/core-common': '7.0.0-beta.29',
-  '@storybook/core-events': '7.0.0-beta.29',
-  '@storybook/core-server': '7.0.0-beta.29',
-  '@storybook/core-webpack': '7.0.0-beta.29',
-  '@storybook/csf-tools': '7.0.0-beta.29',
-  '@storybook/docs-tools': '7.0.0-beta.29',
-  '@storybook/ember': '7.0.0-beta.29',
-  '@storybook/html': '7.0.0-beta.29',
-  '@storybook/html-webpack5': '7.0.0-beta.29',
-  '@storybook/html-vite': '7.0.0-beta.29',
-  '@storybook/instrumenter': '7.0.0-beta.29',
-  '@storybook/nextjs': '7.0.0-beta.29',
-  '@storybook/node-logger': '7.0.0-beta.29',
-  '@storybook/postinstall': '7.0.0-beta.29',
-  '@storybook/preact': '7.0.0-beta.29',
-  '@storybook/preact-vite': '7.0.0-beta.29',
-  '@storybook/preact-webpack5': '7.0.0-beta.29',
-  '@storybook/preset-html-webpack': '7.0.0-beta.29',
-  '@storybook/preset-preact-webpack': '7.0.0-beta.29',
-  '@storybook/preset-react-webpack': '7.0.0-beta.29',
-  '@storybook/preset-server-webpack': '7.0.0-beta.29',
-  '@storybook/preset-svelte-webpack': '7.0.0-beta.29',
-  '@storybook/preset-vue-webpack': '7.0.0-beta.29',
-  '@storybook/preset-vue3-webpack': '7.0.0-beta.29',
-  '@storybook/preset-web-components-webpack': '7.0.0-beta.29',
-  '@storybook/preview': '7.0.0-beta.29',
-  '@storybook/preview-api': '7.0.0-beta.29',
-  '@storybook/preview-web': '7.0.0-beta.29',
-  '@storybook/react': '7.0.0-beta.29',
-  '@storybook/react-vite': '7.0.0-beta.29',
-  '@storybook/react-webpack5': '7.0.0-beta.29',
-  '@storybook/router': '7.0.0-beta.29',
-  '@storybook/server': '7.0.0-beta.29',
-  '@storybook/server-webpack5': '7.0.0-beta.29',
-  '@storybook/source-loader': '7.0.0-beta.29',
-  '@storybook/store': '7.0.0-beta.29',
-  '@storybook/svelte': '7.0.0-beta.29',
-  '@storybook/sveltekit': '7.0.0-beta.29',
-  '@storybook/svelte-vite': '7.0.0-beta.29',
-  '@storybook/svelte-webpack5': '7.0.0-beta.29',
-  '@storybook/telemetry': '7.0.0-beta.29',
-  '@storybook/theming': '7.0.0-beta.29',
-  '@storybook/manager': '7.0.0-beta.29',
-  '@storybook/vue': '7.0.0-beta.29',
-  '@storybook/vue-vite': '7.0.0-beta.29',
-  '@storybook/vue-webpack5': '7.0.0-beta.29',
-  '@storybook/vue3': '7.0.0-beta.29',
-  '@storybook/vue3-vite': '7.0.0-beta.29',
-  '@storybook/vue3-webpack5': '7.0.0-beta.29',
-  '@storybook/web-components': '7.0.0-beta.29',
-  '@storybook/web-components-webpack5': '7.0.0-beta.29',
-  '@storybook/web-components-vite': '7.0.0-beta.29',
-  sb: '7.0.0-beta.29',
-  storybook: '7.0.0-beta.29',
+  '@storybook/addon-a11y': '7.0.0-beta.31',
+  '@storybook/addon-actions': '7.0.0-beta.31',
+  '@storybook/addon-backgrounds': '7.0.0-beta.31',
+  '@storybook/addon-controls': '7.0.0-beta.31',
+  '@storybook/addon-docs': '7.0.0-beta.31',
+  '@storybook/addon-essentials': '7.0.0-beta.31',
+  '@storybook/addon-highlight': '7.0.0-beta.31',
+  '@storybook/addon-interactions': '7.0.0-beta.31',
+  '@storybook/addon-jest': '7.0.0-beta.31',
+  '@storybook/addon-links': '7.0.0-beta.31',
+  '@storybook/addon-measure': '7.0.0-beta.31',
+  '@storybook/addon-outline': '7.0.0-beta.31',
+  '@storybook/addon-storyshots': '7.0.0-beta.31',
+  '@storybook/addon-storyshots-puppeteer': '7.0.0-beta.31',
+  '@storybook/addon-storysource': '7.0.0-beta.31',
+  '@storybook/addon-toolbars': '7.0.0-beta.31',
+  '@storybook/addon-viewport': '7.0.0-beta.31',
+  '@storybook/addons': '7.0.0-beta.31',
+  '@storybook/angular': '7.0.0-beta.31',
+  '@storybook/manager-api': '7.0.0-beta.31',
+  '@storybook/blocks': '7.0.0-beta.31',
+  '@storybook/builder-manager': '7.0.0-beta.31',
+  '@storybook/builder-vite': '7.0.0-beta.31',
+  '@storybook/builder-webpack5': '7.0.0-beta.31',
+  '@storybook/channel-postmessage': '7.0.0-beta.31',
+  '@storybook/channel-websocket': '7.0.0-beta.31',
+  '@storybook/channels': '7.0.0-beta.31',
+  '@storybook/cli': '7.0.0-beta.31',
+  '@storybook/client-api': '7.0.0-beta.31',
+  '@storybook/client-logger': '7.0.0-beta.31',
+  '@storybook/codemod': '7.0.0-beta.31',
+  '@storybook/components': '7.0.0-beta.31',
+  '@storybook/core-client': '7.0.0-beta.31',
+  '@storybook/core-common': '7.0.0-beta.31',
+  '@storybook/core-events': '7.0.0-beta.31',
+  '@storybook/core-server': '7.0.0-beta.31',
+  '@storybook/core-webpack': '7.0.0-beta.31',
+  '@storybook/csf-tools': '7.0.0-beta.31',
+  '@storybook/docs-tools': '7.0.0-beta.31',
+  '@storybook/ember': '7.0.0-beta.31',
+  '@storybook/html': '7.0.0-beta.31',
+  '@storybook/html-webpack5': '7.0.0-beta.31',
+  '@storybook/html-vite': '7.0.0-beta.31',
+  '@storybook/instrumenter': '7.0.0-beta.31',
+  '@storybook/nextjs': '7.0.0-beta.31',
+  '@storybook/node-logger': '7.0.0-beta.31',
+  '@storybook/postinstall': '7.0.0-beta.31',
+  '@storybook/preact': '7.0.0-beta.31',
+  '@storybook/preact-vite': '7.0.0-beta.31',
+  '@storybook/preact-webpack5': '7.0.0-beta.31',
+  '@storybook/preset-html-webpack': '7.0.0-beta.31',
+  '@storybook/preset-preact-webpack': '7.0.0-beta.31',
+  '@storybook/preset-react-webpack': '7.0.0-beta.31',
+  '@storybook/preset-server-webpack': '7.0.0-beta.31',
+  '@storybook/preset-svelte-webpack': '7.0.0-beta.31',
+  '@storybook/preset-vue-webpack': '7.0.0-beta.31',
+  '@storybook/preset-vue3-webpack': '7.0.0-beta.31',
+  '@storybook/preset-web-components-webpack': '7.0.0-beta.31',
+  '@storybook/preview': '7.0.0-beta.31',
+  '@storybook/preview-api': '7.0.0-beta.31',
+  '@storybook/preview-web': '7.0.0-beta.31',
+  '@storybook/react': '7.0.0-beta.31',
+  '@storybook/react-vite': '7.0.0-beta.31',
+  '@storybook/react-webpack5': '7.0.0-beta.31',
+  '@storybook/router': '7.0.0-beta.31',
+  '@storybook/server': '7.0.0-beta.31',
+  '@storybook/server-webpack5': '7.0.0-beta.31',
+  '@storybook/source-loader': '7.0.0-beta.31',
+  '@storybook/store': '7.0.0-beta.31',
+  '@storybook/svelte': '7.0.0-beta.31',
+  '@storybook/sveltekit': '7.0.0-beta.31',
+  '@storybook/svelte-vite': '7.0.0-beta.31',
+  '@storybook/svelte-webpack5': '7.0.0-beta.31',
+  '@storybook/telemetry': '7.0.0-beta.31',
+  '@storybook/theming': '7.0.0-beta.31',
+  '@storybook/manager': '7.0.0-beta.31',
+  '@storybook/vue': '7.0.0-beta.31',
+  '@storybook/vue-vite': '7.0.0-beta.31',
+  '@storybook/vue-webpack5': '7.0.0-beta.31',
+  '@storybook/vue3': '7.0.0-beta.31',
+  '@storybook/vue3-vite': '7.0.0-beta.31',
+  '@storybook/vue3-webpack5': '7.0.0-beta.31',
+  '@storybook/web-components': '7.0.0-beta.31',
+  '@storybook/web-components-webpack5': '7.0.0-beta.31',
+  '@storybook/web-components-vite': '7.0.0-beta.31',
+  sb: '7.0.0-beta.31',
+  storybook: '7.0.0-beta.31',
 };
