@@ -164,8 +164,8 @@ export class StoryIndexGenerator {
     );
   }
 
-  isDocsMdx(absolutePath: Path) {
-    return /(?<!\.stories)\.mdx$/i.test(absolutePath);
+  isStoryFile(absolutePath: Path) {
+    return /(\.(js|jsx|ts|tsx)|\.stories\.mdx)$/i.test(absolutePath);
   }
 
   async ensureExtracted(): Promise<IndexEntry[]> {
@@ -174,7 +174,7 @@ export class StoryIndexGenerator {
     // files may use the `<Meta of={XStories} />` syntax, which requires
     // that the story file that contains the meta be processed first.
     await this.updateExtracted(async (specifier, absolutePath) =>
-      this.isDocsMdx(absolutePath) ? false : this.extractStories(specifier, absolutePath)
+      this.isStoryFile(absolutePath) ? this.extractStories(specifier, absolutePath) : false
     );
 
     if (!this.options.docs.disable) {
