@@ -23,8 +23,8 @@ export async function getPreviewBuilder(
 }
 
 export async function getBuilders({ presets, configDir }: Options): Promise<Builder<unknown>[]> {
-  const core = await presets.apply<CoreConfig>('core', undefined);
-  const builderName = typeof core?.builder === 'string' ? core.builder : core?.builder?.name;
+  const { builder } = await presets.apply<CoreConfig>('core', {});
+  const builderName = typeof builder === 'string' ? builder : builder?.name;
 
   return Promise.all([getPreviewBuilder(builderName, configDir), getManagerBuilder()]);
 }
