@@ -208,10 +208,24 @@ export function popParametersObjectFromDefaultExport(source, ast) {
       patchNode(node);
 
       const isDefaultExport = node.type === 'ExportDefaultDeclaration';
-      const isObjectExpression = node.declaration?.type === 'ObjectExpression';
-      const isTsAsExpression = node.declaration?.type === 'TSAsExpression';
+      let decl = node.declaration;
 
-      const targetNode = isObjectExpression ? node.declaration : node.declaration?.expression;
+      // handle `const meta = { }; export default meta;`
+      if (isDefaultExport && decl?.type === 'Identifier') {
+        ast.body.forEach((n) => {
+          if (n.type === 'VariableDeclaration') {
+            n.declarations.forEach((d) => {
+              if (d.id.name === decl.name) {
+                decl = d.init;
+              }
+            });
+          }
+        });
+      }
+
+      const isObjectExpression = decl?.type === 'ObjectExpression';
+      const isTsAsExpression = decl?.type === 'TSAsExpression';
+      const targetNode = isObjectExpression ? decl : decl?.expression;
 
       if (
         isDefaultExport &&
