@@ -1,13 +1,20 @@
 import React from 'react';
-import type { Meta, StoryObj } from '@storybook/react';
-import type { ModuleExport } from '@storybook/types';
+import type { Meta, ReactRenderer, StoryObj } from '@storybook/react';
+import { within } from '@storybook/testing-library';
+import type { PlayFunctionContext } from '@storybook/csf';
+import type { WebRenderer, ModuleExport } from '@storybook/types';
+import { RESET_STORY_ARGS, STORY_ARGS_UPDATED, UPDATE_STORY_ARGS } from '@storybook/core-events';
+import type { PreviewWeb } from '@storybook/preview-api';
+import type { Channel } from '@storybook/channels';
+
 import type { StoryProps } from './Story';
 import { Story as StoryComponent, StorySkeleton } from './Story';
 import type { DocsContextProps } from '../blocks';
 import * as ButtonStories from '../examples/Button.stories';
 
 // eslint-disable-next-line no-underscore-dangle
-const preview = (window as any).__STORYBOOK_PREVIEW__;
+const preview = (window as any).__STORYBOOK_PREVIEW__ as PreviewWeb<ReactRenderer>;
+const channel = (window as any).__STORYBOOK_ADDONS_CHANNEL__ as Channel;
 const renderStoryToElement = preview.renderStoryToElement.bind(preview);
 
 type ExtendedStoryProps = Omit<StoryProps, 'story'> & {
@@ -22,13 +29,13 @@ const meta: Meta<ExtendedStoryProps> = {
   },
   args: {
     height: '100px',
+    primary: false,
     // NOTE: the real story arg is a PreparedStory, which we'll get in the render function below
     storyExport: ButtonStories.Primary as any,
   },
   render({ storyExport, ...args }, { loaded }) {
     const docsContext = loaded.docsContext as DocsContextProps;
-    const resolved = docsContext.resolveModuleExport(storyExport);
-    if (resolved.type !== 'story') throw new Error('Bad export, pass a story!');
+    const resolved = docsContext.resolveOf(storyExport, ['story']);
     // @ts-expect-error getting too complex with props
     return <StoryComponent {...args} story={resolved.story} />;
   },
@@ -44,6 +51,7 @@ export const Inline: Story = {
     inline: true,
     // @ts-expect-error getting too complex with props
     autoplay: false,
+    forceInitialArgs: false,
     renderStoryToElement,
   },
 };
@@ -52,6 +60,38 @@ export const IFrame: Story = {
   name: 'IFrame',
   args: {
     inline: false,
+  },
+};
+
+export const ForceInitialArgs = {
+  args: {
+    storyExport: ButtonStories.Primary,
+    inline: true,
+    autoplay: true,
+    forceInitialArgs: true,
+    renderStoryToElement,
+  },
+  // test that it ignores updated args by emitting an arg update and assert that it isn't reflected in the DOM
+  play: async ({ args, canvasElement, loaded }: PlayFunctionContext<WebRenderer>) => {
+    const docsContext = loaded.docsContext as DocsContextProps;
+    const resolved = docsContext.resolveOf(args.storyExport, ['story']);
+
+    await within(canvasElement).findByText(/Button/);
+
+    const updatedPromise = new Promise<void>((resolve) => {
+      channel.once(STORY_ARGS_UPDATED, resolve);
+    });
+    await channel.emit(UPDATE_STORY_ARGS, {
+      storyId: resolved.story.id,
+      updatedArgs: { label: 'Updated' },
+    });
+    await updatedPromise;
+    await within(canvasElement).findByText(/Button/);
+
+    await channel.emit(RESET_STORY_ARGS, { storyId: resolved.story.id });
+    await new Promise<void>((resolve) => {
+      channel.once(STORY_ARGS_UPDATED, resolve);
+    });
   },
 };
 
